% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{theme_colour_xipu3}
\alias{theme_colour_xipu3}
\title{Add Xipu theme colour (3 colours) to the legends}
\format{An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}) of length 16.}
\usage{
theme_colour_xipu3
}
\arguments{
\item{NA}{No paramater needed. Use this function with ggplot2}
}
\description{
This function add colour to THREE discrete colours
}
\keyword{ggplot2}
\keyword{theme}
