% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xipu.R
\name{theme_xipu}
\alias{theme_xipu}
\title{Professional plot look for Xipu only}
\usage{
Append the function to ggplot2 statements as you would do for ggthemes,
}
\arguments{
\item{NA}{No paramater needed. Use this function with ggplot2}
}
\description{
This function allows you to plot a high standard plot
}
\note{
Fonts needed: EB Garamond (or equivalent). Birch Std(or equivalent)
}
\keyword{ggplot2}
\keyword{theme}
